Update Video Poker game to determine the poker hand  winner for each round
Update Video Poker game to determine the poker hand  winner for each round.  Need to update the bank roll, etc.  For now, wanted to check in the working code.  Will fix the rest of the code for the design to make sure it matches the requirements

+++ Python/source/ben/Player.py
@@ -75,7 +75,11 @@ class Player(object):
      @return: NONE
      '''
      def remove_funds(self, amount_of_funds):
-        self.__bank_roll = self.__bank_roll - int(amount_of_funds)
+         
+        # probably should check to make sure the player as enough funds in their 
+        # bank roll before blindly deduction cash from the account.
+        # we dont want our account to go in the red or become negative
+        self.__bank_roll = self.__bank_roll - float(amount_of_funds)
     
      '''
      get_funds
